extracted commond code to RepresentativeService
